+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release Notes
 
+## v1.0.7 (2022-06-22)
+
+- Fixed appearing transition with background images. Now it works.
+- Improved documentation `README.md` file.
+- Added **configurations** for smooth loading images.
+
+----
+
 ## v1.0.6 (2022-01-2)
 
 - Changed default images selector to be more specific.